requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
pandas==2.2.1
openpyxl==3.1.2
requests==2.31.0
pillow==10.2.0
python-multipart==0.0.9